* Added: [Debian] multi-arch tag. * Added: [Debian] To compile using all installed CPUs. * Modified: Editor desktop files to allow 32 -> 64 bits migration. * Added: Support for LP64 architectures (i.e. AMD64) * Fixed: Uses of memcpy that could potentially generate problems, some of them were problematic. * Added: More checks when reading setenvir.dat. * Fixed: [txhgen.cc] Potential use of unitilialized var. * Fixed: [Errors parser] to support the new gnu make quotation style.

+++ setedit/setedit/dstfile.cc
@@ -1,4 +1,4 @@
-/* Copyright (C) 1996-2007 by Salvador E. Tropea (SET),
+/* Copyright (C) 1996-2016 by Salvador E. Tropea (SET),
    see copyrigh file for details */
 //#define DEBUG
 #include <ceditint.h>
@@ -653,7 +653,7 @@ void TSetEditorApp::storeDesktop(fpstream& s)
  // Save a signature to identify the file
  s.writeString(Signature);
  // Save the version
- s << TCEDITOR_VERSION;
+ s << (uint32)TCEDITOR_VERSION;
 
  // Save the video mode & font, first to avoid a lot of redraw
  if (soCol)
